catalogue-service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
python-multipart
python-jose[cryptography]